Large box shadow is not clipped by clip-path() when in scroll container https://bugs.webkit.org/show_bug.cgi?id=282106 rdar://138599295
Reviewed by NOBODY (OOPS!).

When a layer with a mask toggled between tiled and non-tiled, we correctly call `GraphicsLayerCA::updateMaskLayer()`
on the WebCore side, but on the WebKit side we hit a quirk of CoreAnimation, which is that assigning a CALayer's mask
layer sets the superlayer of the mask layer. If the superlayer was non-nil, we'd assert and early return, rather than
assigning the mask layer to its new owner.

* LayoutTests/compositing/masks/clip-path-on-tiled-toggle-expected.html: Added.
* LayoutTests/compositing/masks/clip-path-on-tiled-toggle.html: Added.
* Source/WebKit/Shared/RemoteLayerTree/RemoteLayerTreePropertyApplier.mm:
(WebKit::RemoteLayerTreePropertyApplier::updateMask):