Interrupting scroll momentum causes container to not render https://bugs.webkit.org/show_bug.cgi?id=262287 rdar://116205365
Reviewed by NOBODY (OOPS!).

On macOS, when scroll momentum is interrupted with a programmatic scroll, it's possible for the
web process to end up with a stale scroll position, which results in missing tiles.

This happens when the IPC for the programmatic scroll from the web to the UI process overlaps
with the IPC for the scroll resulting from a wheel event; the messages pass like ships in the night.

Fix by not early returning from `RemoteScrollingTreeMac::scrollingTreeNodeDidScroll()`, so that
we send a message back to the web process in response to the programmatic scroll.

* LayoutTests/fast/scrolling/mac/momentum-then-programmatic-missing-tiles-expected.html: Added.
* LayoutTests/fast/scrolling/mac/momentum-then-programmatic-missing-tiles.html: Added.
* Source/WebKit/UIProcess/RemoteLayerTree/mac/RemoteScrollingTreeMac.mm:
(WebKit::RemoteScrollingTreeMac::scrollingTreeNodeDidScroll):