Tabbing in a scroll container with scroll-padding set does not scroll focused element into view https://bugs.webkit.org/show_bug.cgi?id=290096 rdar://147513379
Reviewed by NOBODY (OOPS!).

The existing code took scroll-padding into account by inflating the `localExposeRect`, but this doesn't
have the correct behavior because it makes this rect wider. The correct approach is to shrink the
the scrolling viewport. We can also simplify the code by keeping all the rects in coordinates local
to the RenderBox, so `paddingBoxRect()` gives the correct scroll viewport with the left scrollbar
taken into account.

Also handle the case of the scroll padding being larger than the scrollport by proportionally shrinking
it; this behavior is not specified, but seems reasonable.

* LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-padding-focus-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-padding-focus.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-padding-percentage-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-padding-percentage.html: Added.
* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebCore/rendering/RenderLayerScrollableArea.cpp:
(WebCore::computeLocalExposeRect):
(WebCore::RenderLayerScrollableArea::scrollRectToVisible):
(WebCore::RenderLayerScrollableArea::updateScrollPositionForScrollIntoView): This only needs the offset to
reveal, not the entire rect.
(WebCore::getLocalExposeRect): Deleted.
* Source/WebCore/rendering/RenderLayerScrollableArea.h: